+++ Content.Shared/Chemistry/Reagent/ReagentPrototype.cs
@@ -51,7 +51,7 @@ namespace Content.Shared.Chemistry.Reagent
         ///     How much energy it takes to heat one unit of this reagent by one Kelvin.
         /// </summary>
         [DataField("specificHeat")]
-        public double SpecificHeat { get; } = 1.0d;
+        public float SpecificHeat { get; } = 1.0f;
 
         [DataField("boilingPoint")]
         public float? BoilingPoint { get; }
